muda para uint8

+++ aux.c
@@ -1,29 +1,30 @@
 #include <stdio.h>
 #include <stdlib.h>
+#include <stdint.h>
 #include "aux.h"
 
-int **alocaMatriz(int m, int n) {
-    int **tabuleiro = (int **)malloc(m * sizeof(int *));
-    for (int i = 0; i < m; i++) 
-        tabuleiro[i] = (int *)malloc(n * sizeof(int));
+uint8_t **alocaMatriz(int m, int n) {
+    uint8_t **tabuleiro = (uint8_t **)malloc(m * sizeof(uint8_t *));
+    for (uint8_t i = 0; i < m; i++) 
+        tabuleiro[i] = (uint8_t *)malloc(n * sizeof(uint8_t));
     return tabuleiro;
 }
 
-void read_tabuleiro(int **tabuleiro, int m, int n) {
+void read_tabuleiro(uint8_t **tabuleiro, int m, int n) {
     for (int i = 0; i < m; i++) 
         for (int j = 0; j < n; j++) 
-            scanf("%d", &tabuleiro[i][j]);
+            scanf("%hhd", &tabuleiro[i][j]);
 }
 
-void print_tabuleiro(int **tabuleiro, int m, int n) {
+void print_tabuleiro(uint8_t **tabuleiro, int m, int n) {
     for (int i = 0; i < m; i++) {
         for (int j = 0; j < n; j++) 
-            printf("%d ", tabuleiro[i][j]);
+            printf("%hhd ", tabuleiro[i][j]);
         printf("\n");
     }
 }
 
-void destroy_tabuleiro(int **tabuleiro, int m) {
+void destroy_tabuleiro(uint8_t **tabuleiro, int m) {
     for (int i = 0; i < m; i++) 
         free(tabuleiro[i]);
     free(tabuleiro);
